requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
greenlet==3.2.4
idna==3.10
mysql-connector-python==9.5.0
playwright==1.55.0
pyee==13.0.0
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
requests==2.32.5
telebot==0.0.5
typing_extensions==4.15.0
urllib3==2.5.0
